+++ docs/changelog.md
@@ -13,7 +13,8 @@ category: Administration
 
 * feat: ensure compatibility with Jira 9.10.0, 9.4.8 and 8.20.24
 * feat: create subtasks with custom fields in single transaction
-* feat: improve display of errors not related to subtask syntax 
+* feat: improve display of errors not related to subtask syntax
+* fix: hide dialog button in issues without subtasks available (#89)
 
 ### [23.07.0] - 2023-07-02